pyproject.toml
[tool.poetry]
name = "make-animeface-dataset"
version = "0.1.0"
description = ""
authors = ["saityyy <62131533+saityyy@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "make_animeface_dataset"}]

[tool.poetry.dependencies]
python = "^3.9"
flake8 = "^6.0.0"
autopep8 = "^2.0.2"
pillow = "^10.0.0"
tqdm = "^4.65.0"
numpy = "^1.25.1"
opencv-python = "^4.8.0.74"
matplotlib = "^3.7.2"
torch = "^2.0.1+cu118"
torchvision = "^0.15.2+cu118"
pyyaml = "^6.0"
pandas = "^2.0.3"


[[tool.poetry.source]]
name = "torch_cu118"
url = "https://download.pytorch.org/whl/cu118"
priority = "primary"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
